refactor(types): split webhooks.ts into entities/{webhook-config,webhook-delivery}.ts
