+++ packages/types/src/entities/webhook-config.ts
@@ -1,5 +1,4 @@
-import type { ServerSecret } from "./encryption.js";
-import type { BucketContentEntityType } from "./entities/bucket.js";
+import type { ServerSecret } from "../encryption.js";
 import type {
   AccountId,
   AcknowledgementId,
@@ -19,14 +18,10 @@ import type {
   PollId,
   PollVoteId,
   SystemId,
-  WebhookDeliveryId,
   WebhookId,
-} from "./ids.js";
-import type { UnixMillis } from "./timestamps.js";
-import type { Archived, AuditMetadata } from "./utility.js";
-
-/** Status of a webhook delivery attempt. */
-export type WebhookDeliveryStatus = "pending" | "success" | "failed";
+} from "../ids.js";
+import type { Archived, AuditMetadata } from "../utility.js";
+import type { BucketContentEntityType } from "./bucket.js";
 
 /** Events that can trigger a webhook. */
 export type WebhookEventType =
@@ -265,17 +260,3 @@ export interface WebhookConfig extends AuditMetadata {
 
 /** An archived webhook config. */
 export type ArchivedWebhookConfig = Archived<WebhookConfig>;
-
-/** A record of a webhook delivery attempt with retry lifecycle. */
-export interface WebhookDelivery {
-  readonly id: WebhookDeliveryId;
-  readonly systemId: SystemId;
-  readonly webhookId: WebhookId;
-  readonly eventType: WebhookEventType;
-  readonly status: WebhookDeliveryStatus;
-  readonly httpStatus: number | null;
-  readonly attemptCount: number;
-  readonly lastAttemptAt: UnixMillis | null;
-  readonly nextRetryAt: UnixMillis | null;
-  readonly createdAt: UnixMillis;
-}
